+++ lib/pages/chat/events/redacted_content.dart
@@ -42,7 +42,7 @@ class EventRedactedContent extends StatelessWidget {
                     ? 'SystemFont'
                     : AppSettings.chatFont.value
               : null,
-          fontFamilyFallback: AppSettings.chatFallbackFonts.value.split(','),
+          fontFamilyFallback: AppSettings.fontFallback(AppSettings.chatFallbackFonts),
         );
 
         return Padding(