+++ lib/pages/chat/trust_user_key_dialog.dart
@@ -102,10 +102,9 @@ Future<bool> showTrustUserInRoomDialog(BuildContext context, Room room) async {
                                       '???',
                                   style: TextStyle(
                                     fontFamily: AppSettings.monospaceFont.value,
-                                    fontFamilyFallback: AppSettings
-                                        .monospaceFallbackFonts
-                                        .value
-                                        .split(','),
+                                    fontFamilyFallback: AppSettings.fontFallback(
+                                      AppSettings.monospaceFallbackFonts,
+                                    ),
                                   ),
                                 ),
                               ),