+++ lib/pages/chat_encryption_settings/chat_encryption_settings_view.dart
@@ -303,11 +303,10 @@ class _ChatEncryptionSettingsViewState
                                                               AppSettings
                                                                   .monospaceFont
                                                                   .value,
-                                                          fontFamilyFallback:
-                                                              AppSettings
-                                                                  .monospaceFallbackFonts
-                                                                  .value
-                                                                  .split(','),
+                                                          fontFamilyFallback: AppSettings.fontFallback(
+                                                            AppSettings
+                                                                .monospaceFallbackFonts,
+                                                          ),
                                                         ),
                                                       ),
                                                       leading: CircleAvatar(