+++ lib/pages/settings_homeserver/settings_homeserver_view.dart
@@ -294,10 +294,9 @@ class SettingsHomeserverView extends StatelessWidget {
                                   style: TextStyle(
                                     color: theme.colorScheme.onSurface,
                                     fontFamily: AppSettings.monospaceFont.value,
-                                    fontFamilyFallback: AppSettings
-                                        .monospaceFallbackFonts
-                                        .value
-                                        .split(','),
+                                    fontFamilyFallback: AppSettings.fontFallback(
+                                      AppSettings.monospaceFallbackFonts,
+                                    ),
                                   ),
                                 ),
                               ),